Rename widget from the `MetricsSwitch` to the `Toggle`.
