+++ metrics/web/test/common/presentation/toggle/theme/toggle_theme_data_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:metrics/common/presentation/switch/theme/toggle_theme_data.dart';
+import 'package:metrics/common/presentation/toggle/theme/toggle_theme_data.dart';
 import 'package:test/test.dart';
 
 // https://github.com/software-platform/monorepo/issues/140
